packaging: Reduce rebuilds with accurate source filter
This removes `srcDir`, which is not needed. It is ok to refer to local
files.

The filter is "positive" or inclusion-based, because exclusion-based
filters tend to include too much, causing unnecessary rebuilds, and
they require more maintenance as files are added. Sometimes it's a
trade-off with list size, but that's not even the case here.

This commit assumes the `default.nix` is only accessed through flakes.
Otherwise, it'd be good to filter out `build` again.
I have not implemented that, because I found no other traces of
non-flake usage.